# services/api/openapi.proposal.yaml
...
openapi: 3.1.0
info:
  title: TuneDeck API
  version: 0.2.0-proposal
servers:
  - url: /
paths:
...
  /v1/admin/jobs:
    get:
      summary: Background jobs of every queue, with per-queue counts (operator, admin)
      description: 'Roles: operator or admin. Queues account_deletion, account_export and (only when a Keycloak admin client is configured) idp_session_end. Dead letters first, then retrying, then by request time; at most 100 jobs (truncated true when there are more). Completed jobs are listed for 7 days. Ids are opaque.'
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, failed, dead_letter, completed, all], default: open }, description: 'open = pending, retrying and dead_letter; failed = retrying and dead_letter' }
      responses:
        '200':
          description: Jobs
          headers: { Cache-Control: { schema: { type: string, example: no-store } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminJobList' } } }
        '400': { description: 'VALIDATION_FAILED: field status, reason invalid', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED, or MFA_REQUIRED with details {scope: session, maxAgeSeconds: 43200}', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ROLE_REQUIRED: needs operator or admin', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/admin/jobs/{jobId}/retry:
    parameters:
      - { name: jobId, in: path, required: true, schema: { type: string, pattern: '^(?:[0-9a-f]{64}|ex_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|se_[0-9a-f]{64})$' }, description: 'A deletion ticket hash, ex_<export id> or se_<sha256>; anything else is 404' }
    post:
      summary: Run one job now (operator, admin)
      description: 'Roles: operator or admin. Deletions run under the queue lock; a dead-lettered job starts a new round of 5 attempts. A job that already finished returns completed and runs nothing (not audited). At most 3 staff retries per job in 24 hours, counted from the audit trail. Audited as job.retry with the reason, kind, state and attempts before, and the result.'
      parameters:
        - { name: Idempotency-Key, in: header, required: false, description: 'Optional. A repeat with the same body within 24 hours gets the first response again (Idempotent-Replayed: true)', schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminExportReason' } } } }
      responses:
        '200':
          description: The job's status afterwards
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [pending, retrying, dead_letter, completed] } } } } }
        '400': { description: 'VALIDATION_FAILED: reason (required, length, control_characters)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED, or MFA_REQUIRED with details {scope: session, maxAgeSeconds: 43200}', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ROLE_REQUIRED: needs operator or admin', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 'NOT_FOUND: unknown or malformed id, an idp_session_end job without a Keycloak admin client, or an export that is gone', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 'JOB_RETRY_LIMIT with details {limit: 3, windowHours: 24}; JOB_BUSY while a background run holds the job; JOB_SUPERSEDED for a dead-lettered export whose account asked for a new one; or IDEMPOTENCY_KEY_REUSED / IDEMPOTENCY_IN_PROGRESS', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { $ref: '#/components/responses/Error' }
        '429': { description: 'JOB_RETRY_TOO_SOON within 60 s of the last attempt, with details.retryAfterSeconds and the Retry-After header (the seconds left)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT, description: 'OIDC access token (iss/aud/exp/RS256 verified via JWKS)' }
...
  schemas:
...
    Error:
      type: object
      required: [code, messageKey, requestId, details]
      properties:
        code:
          type: string
          enum: [AUTH_REQUIRED, AUTH_FORBIDDEN, VALIDATION_FAILED, NOT_FOUND, EXPORT_NOT_READY, IDEMPOTENCY_KEY_REQUIRED, PRECONDITION_REQUIRED, REVISION_MISMATCH, PAYLOAD_TOO_LARGE, DEPENDENCY_UNAVAILABLE, REAUTH_REQUIRED, DEVICE_REVOKED, DEVICE_LIMIT, ROLE_REQUIRED, PUBLISH_BLOCKED, CHECK_TOO_SOON, CHECKER_UNAVAILABLE, RIGHTS_ALREADY_REVOKED, API_RATE_LIMITED, ACCOUNT_DELETING, JOB_RETRY_TOO_SOON, JOB_BUSY, JOB_RETRY_LIMIT, JOB_SUPERSEDED, SYNC_RESET_REQUIRED, PURCHASE_INVALID, PURCHASE_CONFLICT, MFA_REQUIRED, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, SUPPORT_ACCESS_REQUIRED, INTERNAL, DIRECTORY_BLOCK_EXISTS]
        messageKey: { type: string, example: errors.settings.revisionMismatch }
        requestId: { type: string }
        details:
          type: object
          description: 'No tokens or PII. VALIDATION_FAILED → {field?, reason, allowed?}; REVISION_MISMATCH → {currentRevision}; REAUTH_REQUIRED → {maxAgeSeconds}; MFA_REQUIRED → {maxAgeSeconds, scope?}; DEVICE_LIMIT → {maxActiveDevices}; PUBLISH_BLOCKED → {reasons}; IDEMPOTENCY_IN_PROGRESS / CHECK_TOO_SOON / JOB_RETRY_TOO_SOON → {retryAfterSeconds}.'
...
    AdminExportReason:
      type: object
      required: [reason]
      description: 'Reason for an export, a job retry and similar recorded actions. Trimmed, then 10 to 500 characters with no control characters. Other fields are ignored.'
      properties:
        reason: { type: string, minLength: 10, maxLength: 500 }
...
    AdminJob:
      type: object
      required: [id, kind, status, requestedAt, attempts, maxAttempts, lastAttemptAt, nextAttemptAt, lastErrorCode, completedAt, deadline]
      properties:
        id: { type: string, description: 'Opaque: a deletion ticket hash, ex_<export id> or se_<sha256>; never a user, device or Keycloak session id' }
        kind: { type: string, enum: [account_deletion, account_export, idp_session_end] }
        status: { type: string, enum: [pending, retrying, dead_letter, completed] }
        requestedAt: { type: string, format: date-time }
        attempts: { type: integer, description: Attempts in the current round }
        maxAttempts: { type: integer, const: 5 }
        lastAttemptAt: { type: [string, 'null'], format: date-time }
        nextAttemptAt: { type: [string, 'null'], format: date-time, description: 'Null when dead-lettered, completed or not yet scheduled' }
        lastErrorCode: { type: [string, 'null'], description: 'A code only, e.g. IDP_DELETE_FAILED, DB_40001' }
        completedAt: { type: [string, 'null'], format: date-time }
        deadline: { type: [string, 'null'], format: date-time, description: Account deletions only (30 days from the request) }
    AdminJobList:
      type: object
      required: [jobs, counts, queues, truncated]
      properties:
        jobs: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/AdminJob' } }
        counts:
          type: object
          description: All listed jobs by status, whatever the filter
          required: [pending, retrying, dead_letter, completed]
          properties: { pending: { type: integer }, retrying: { type: integer }, dead_letter: { type: integer }, completed: { type: integer } }
        queues:
          type: array
          items:
            type: object
            required: [kind, pending, retrying, deadLetter, oldestOpenAt]
            properties:
              kind: { type: string, enum: [account_deletion, account_export, idp_session_end] }
              pending: { type: integer }
              retrying: { type: integer }
              deadLetter: { type: integer }
              oldestOpenAt: { type: [string, 'null'], format: date-time }
        truncated: { type: boolean, description: More than 100 jobs matched }
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }